从文件读取分数是坏的，别用！ 救命啊，这读文件真的好麻烦啊

+++ 没有人比我更懂石头剪刀布/player.cpp
@@ -32,10 +32,10 @@ auto twd_player::choose()->int {
 
 twd_player::twd_player() {
 	this->name = make_unique<string>("");
-	this->score = make_unique<int>(0);
+	this->score = make_unique<long long>(0);
 	this->choice = make_shared<int[5]>(0);}
 
 twd_player::twd_player(string name) {
 	this->name = make_unique<string>(name);
-	this->score = make_unique<int>(0);
+	this->score = make_unique<long long>(0);
 	this->choice = make_shared<int[5]>(0);}
