+++ 没有人比我更懂石头剪刀布/player.h
@@ -4,7 +4,7 @@
 class twd_player{
 public:
 	unique_ptr<string> name;
-	unique_ptr<int> score;
+	unique_ptr<long long> score;
 	shared_ptr<int[]> choice;
 	virtual auto choose()->int;
 	twd_player();
